requirements.txt
pytest
pytest-html
pytest-xdist
selenium
allure-pytest